+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# 0.15.0 (2022-xx-xx)
+# 0.14.1 (2022-05-xx)
 
 ## Update
 
@@ -12,10 +12,11 @@
 - [e2e] Testcafe is failing [#750](https://github.com/thorchain/asgardex-electron/issues/750)
 - [Swap] Consider fees in swap limit [#2243](https://github.com/thorchain/asgardex-electron/pull/2242)
 - [Swap] Recipient address lost [#2244](https://github.com/thorchain/asgardex-electron/issues/2244)
+- [Swap] Recipient Ledger address lost [#2252](https://github.com/thorchain/asgardex-electron/issues/2252)
 
 ## Internal
 
-- Update npm dependencies (04-25-2022) - (incl. Electron@18.x, React 18.x, latest xchain-\*, Node 16.13.x etc.) [#2228](https://github.com/thorchain/asgardex-electron/issues/2228)
+- Update npm dependencies (04-25-2022) - (incl. Electron@18.x, React 18.x, latest xchain-\* etc.) [#2228](https://github.com/thorchain/asgardex-electron/issues/2228)
 
 # 0.14.0 (2022-04-23)